refactor: auto-split test_swing_bridge.py
